refactor(market): Wave 1 — move store/queries/mutations to feature module
- marketStore: removed 24 orphaned fields (45→21), moved to features/market/store/
- 5 query files moved to features/market/queries/ (ipos, trending, priceHist, watchlist, offers)
- Trading mutations moved to features/market/mutations/
- Re-export bridges at all old locations for backward compatibility
- tsc 0 errors
