+++ src/features/market/mutations/trading.ts
@@ -1,0 +1,79 @@
+'use client';
+
+import { useMutation } from '@tanstack/react-query';
+import { buyFromMarket, placeBuyOrder, cancelBuyOrder } from '@/lib/services/trading';
+import { buyFromIpo } from '@/lib/services/ipo';
+import { useWallet } from '@/components/providers/WalletProvider';
+import { invalidateTradeQueries } from '@/lib/queries/invalidation';
+import { queryClient } from '@/lib/queryClient';
+import { qk } from '@/lib/queries/keys';
+
+export function useBuyFromMarket() {
+  const { setBalanceCents, refreshBalance } = useWallet();
+  return useMutation({
+    mutationFn: async ({ userId, playerId, quantity }: { userId: string; playerId: string; quantity: number }) => {
+      const result = await buyFromMarket(userId, playerId, quantity);
+      if (!result.success) throw new Error(result.error || 'Kauf fehlgeschlagen');
+      return result;
+    },
+    onSuccess: (result, { playerId, userId }) => {
+      if (result.new_balance != null) setBalanceCents(result.new_balance);
+      refreshBalance();
+      invalidateTradeQueries(playerId, userId);
+      queryClient.invalidateQueries({ queryKey: qk.offers.incoming(userId) });
+    },
+  });
+}
+
+export function useBuyFromIpo() {
+  const { setBalanceCents, refreshBalance } = useWallet();
+  return useMutation({
+    mutationFn: async ({ userId, ipoId, playerId, quantity }: { userId: string; ipoId: string; playerId: string; quantity: number }) => {
+      const result = await buyFromIpo(userId, ipoId, quantity, playerId);
+      if (!result.success) throw new Error(result.error || 'IPO-Kauf fehlgeschlagen');
+      return result;
+    },
+    onSuccess: (result, { playerId, userId }) => {
+      if (result.new_balance != null) setBalanceCents(result.new_balance);
+      refreshBalance();
+      invalidateTradeQueries(playerId, userId);
+      queryClient.invalidateQueries({ queryKey: qk.ipos.active });
+      queryClient.invalidateQueries({ queryKey: qk.ipos.announced });
+      queryClient.invalidateQueries({ queryKey: qk.ipos.recentlyEnded });
+    },
+  });
+}
+
+export function usePlaceBuyOrder() {
+  const { refreshBalance } = useWallet();
+  return useMutation({
+    mutationFn: async ({ userId, playerId, quantity, maxPriceCents }: {
+      userId: string; playerId: string; quantity: number; maxPriceCents: number;
+    }) => {
+      const result = await placeBuyOrder(userId, playerId, quantity, maxPriceCents);
+      if (!result.success) throw new Error(result.error || 'Kaufgesuch fehlgeschlagen');
+      return result;
+    },
+    onSuccess: (_result, { playerId, userId }) => {
+      refreshBalance();
+      invalidateTradeQueries(playerId, userId);
+      queryClient.invalidateQueries({ queryKey: qk.orders.all });
+    },
+  });
+}
+
+export function useCancelBuyOrder() {
+  const { refreshBalance } = useWallet();
+  return useMutation({
+    mutationFn: async ({ userId, orderId }: { userId: string; orderId: string }) => {
+      const result = await cancelBuyOrder(userId, orderId);
+      if (!result.success) throw new Error(result.error || 'Stornierung fehlgeschlagen');
+      return result;
+    },
+    onSuccess: (_result, { userId }) => {
+      refreshBalance();
+      invalidateTradeQueries('', userId);
+      queryClient.invalidateQueries({ queryKey: qk.orders.all });
+    },
+  });
+}
